kumbh-kawach-ai-server/requirements.txt
fastapi
pymongo
google-cloud-storage
google-auth
torch
torchvision
facenet-pytorch
pillow>=11.0
uvicorn[standard]
